Refactor benchmarks
In particular, make the benchmark loop work, so that we can run with
smaller data sizes that need to go through the loop several times.

Also, for the Delta benchmark, don't rely on a potentially big buffer to
store the data. Instead re-generate it as needed by using the same
random seed.