Bugfix: use signed arithmetic in Rollsum.Rotate()
Rollsum.Rotate() was previously doing subtraction on values of type
byte, which in Go always result in a byte (and therefore unsigned)
value. That subtraction was supposed to be signed tough.

This also add unit tests to Rollsum, including a simple one that used to
trigger the bug described above. All "expected" values used in these
unit tests were generated using the original librsync.

Change-type: minor
Signed-off-by: Leandro Motta Barros <leandro@balena.io>