Add patch tests, remove pure delta tests
By "pure delta tests" I mean tests that generate a delta and then
compare it with the reference delta generated by the original rdiff (the
one in C). Generating exactly the same deltas as the original rdiff is
not a worthwhile goal, because this depends on internal details like the
size of output buffers.

The new tests guarantee that our patching is compatible with the deltas
generated by the original rdiff, and that our own delta+patch works as
intended (that is, they end up producing the correct "new file" from the
"old file" and the signature).