Bugfix: bad delta if block is larger than new file
If the new file is smaller than the block size, then a correct delta
shall consist of single instruction:

1. A COPY, if the new file matches the last entry on the signature (the
   one generated from the tail of the old file, whose size is smaller
   than the block size).
2. A LITERAL otherwise (that includes the whole contents of the new
   file).

Previously, we were writing a bogus zero byte to our circular buffer
right on the beginning of the delta generation. This would break the
cases above:

1. It was preventing a proper match with the signature, thus we would
   never generate a COPY command on this case.
2. Our LITERAL would include that bogus zero -- in other words, we would
   generate a corrupted new file.

The main change of this commit is the removal of that bogus
`block.WriteByte(0)`.

The commit also removes the `pos` variable which was not used anywhere.

It also enables all test cases (since now they all pass) and includes
one additional test case that tests the case 1 above.

Change-type: minor
Signed-off-by: Leandro Motta Barros <leandro@balena.io>